+++ rpmio/rpmmessages.h
@@ -70,8 +70,8 @@ extern "C" {
 typedef void * (*rpmCallbackFunction)
 		(/*@null@*/ const void * h, 
 		const rpmCallbackType what, 
-		const unsigned long long amount, 
-		const unsigned long long total,
+		const uint64_t amount, 
+		const uint64_t total,
 		/*@null@*/ fnpyKey key,
 		/*@null@*/ rpmCallbackData data)
 	/*@globals internalState@*/
